Ball -> Bullet, Bat -> Tank
Renaming
